feat(job): implement createPurchase method in JobService and refactor purchase creation logic
