--- web-app/src/lib/services/job/third-party.ts
@@ -1,58 +1,0 @@
-import "server-only";
-
-import { getEnvPublicConfig } from "@/config/env.public";
-import {
-  postPurchase,
-  PostPurchaseResponse,
-} from "@/lib/api/generated/payment";
-import { getPaymentClient } from "@/lib/api/payment-service.client";
-import { AgentWithRelations } from "@/lib/db";
-import { JobInputData } from "@/lib/job-input";
-import { StartJobResponseSchemaType } from "@/lib/schemas";
-import { Err, Ok, Result } from "@/lib/ts-res";
-
-export async function createPurchase(
-  agent: AgentWithRelations,
-  startJobResponse: StartJobResponseSchemaType,
-  inputData: JobInputData,
-  inputHash: string,
-  identifierFromPurchaser: string,
-): Promise<Result<PostPurchaseResponse, string>> {
-  try {
-    const paymentClient = getPaymentClient();
-
-    const postPurchaseResponse = await postPurchase({
-      client: paymentClient,
-      body: {
-        agentIdentifier: agent.blockchainIdentifier,
-        inputHash: inputHash,
-        blockchainIdentifier: startJobResponse.blockchainIdentifier,
-        network: getEnvPublicConfig().NEXT_PUBLIC_NETWORK,
-        sellerVkey: startJobResponse.sellerVKey,
-        paymentType: "Web3CardanoV1",
-        identifierFromPurchaser,
-        payByTime: startJobResponse.payByTime.toString(),
-        externalDisputeUnlockTime:
-          startJobResponse.externalDisputeUnlockTime.toString(),
-        submitResultTime: startJobResponse.submitResultTime.toString(),
-        unlockTime: startJobResponse.unlockTime.toString(),
-        metadata: JSON.stringify({
-          inputData: Object.fromEntries(inputData),
-          jobId: startJobResponse.job_id,
-        }),
-      },
-    });
-
-    if (postPurchaseResponse.error || !postPurchaseResponse.data) {
-      console.log(
-        "Failed to create purchase request",
-        postPurchaseResponse.error,
-      );
-      return Err("Failed to create purchase request");
-    }
-
-    return Ok(postPurchaseResponse.data);
-  } catch (err) {
-    return Err(String(err));
-  }
-}
